Minor changes
Changed error icon to red and warning icon to yellow to improve the
application's appearance. Improved the "Error" function so you can now
customize the error message. Added a settings menu (Nothing in it yet.).
Added code to create a settings.cfg file upon startup if one doesn't
exist, and error if creating/reading it fails.