Minor bug fixes
The resource list's size now resets upon making it in-visible, then
visible again. This means that shrinking it to the point where it's too
small no longer forces you to restart the application in order to made
the resource list useable again.

Made the application "realize" it has to save after adding a sprite.