Updated lots of things!
Updated lots of things including the project license, the readme, the
addition of the LOVE 2D engine, basic settings.cfg file generation
including version number and path to love 2d engine(If you downloaded
the previous commit first, please delete the previously-made
settings.cfg before running this version. Thanks! :D), as well as the
addition of basic "Assembly Information" indicating the program's name,
description, and version number (Which is currently based off of the
number of commits added to the master branch.).