Added ability to delete resources.
Added ability to delete resources by clicking "Edit->Delete Resource" or
pressing the delete key on the keyboard. Had to re-code for quite a
while to get it working correctly, and changed a lot in the process. For
example, the treeview now refreshes after you delete a resource, which
accidentially broke the intended ability to simply keep mashing "del" to
delete everything in order from top to bottom. Now the "Sprites" node is
auto-selected every time you delete a resource. I intend to fix this.